feat: add solution module and hash
